requirements.txt
# Dependências principais do DataMaster SentimentalReview
# ====================================================================

# ====================================================================
# APACHE AIRFLOW E PROVIDERS
# ====================================================================
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-http==4.6.0
apache-airflow-providers-celery==3.4.1
apache-airflow-providers-redis==3.4.0
apache-airflow-providers-apache-spark==4.1.3

# ====================================================================
# PROCESSAMENTO DE DADOS E SPARK
# ====================================================================
pyspark==3.5.0
pandas==2.1.4
numpy==1.24.3
pyarrow==14.0.2
findspark==2.0.1



# ====================================================================
# KAFKA E STREAMING
# ====================================================================
kafka-python==2.0.2
confluent-kafka==2.3.0
avro-python3==1.10.2

# ====================================================================
# MINIO E OBJECT STORAGE
# ====================================================================
minio==7.2.0
boto3==1.34.0
s3fs==2023.12.2
botocore==1.34.0

# ====================================================================
# APIS EXTERNAS
# ====================================================================
python-dotenv==1.0.1
googlemaps==4.10.0
#transformers
#torch==2.2.2+cpu
#--extra-index-url https://download.pytorch.org/whl/cpu
#presidio-analyzer
spacy
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1



# ====================================================================
# BANCO DE DADOS E CONECTORES
# ====================================================================
psycopg2-binary==2.9.9
sqlalchemy==1.4.48
trino==0.328.0
presto-python-client==0.8.4

# ====================================================================
# UTILITÁRIOS CORE
# ====================================================================
python-dotenv==1.0.0
pydantic==2.5.2
loguru==0.7.2
tenacity==8.2.3
click==8.1.7
rich==13.7.0

# ====================================================================
# PROCESSAMENTO DE TEXTO 
# ====================================================================
#nltk==3.8.1
#spacy==3.7.2
#textblob==0.17.1
langdetect==1.0.9

# ====================================================================
# SERIALIZAÇÃO E VALIDAÇÃO
# ====================================================================
jsonschema==4.20.0
marshmallow==3.20.2
pyyaml==6.0.1
toml==0.10.2

# ====================================================================
# DATA E TEMPO
# ====================================================================
pendulum==2.1.2
python-dateutil==2.8.2
pytz==2023.3

# ====================================================================
# NETWORKING E HTTP
# ====================================================================
urllib3==2.1.0
certifi==2023.11.17

# ====================================================================
# CONFIGURAÇÃO ESPECÍFICA 
# ====================================================================
# Dependências específicas para integração MinIO
py4j==0.10.9.7
delta-spark==3.0.0

# ====================================================================
# UTILITÁRIOS DE SISTEMA
# ====================================================================
psutil==5.9.6
memory-profiler==0.61.0

# ====================================================================
# CACHE E PERFORMANCE
# ====================================================================
redis==5.0.1
diskcache==5.6.3

# ====================================================================
# CONFIGURAÇÕES ESPECÍFICAS PARA PRODUÇÃO
# ====================================================================
# Versões fixas para garantir reprodutibilidade
setuptools==69.0.3
wheel==0.42.0
pip==23.3.2

# ====================================================================
# UTILITÁRIOS ADICIONAIS
# ====================================================================
schedule==1.2.0
retrying==1.3.4
tqdm==4.66.1
colorama==0.4.6
tabulate==0.9.0

